Move filters-related search tests to a separate file
